backend/requirements.txt
flask==3.0.0
functions-framework==3.5.0
google-generativeai==0.3.2
gunicorn==21.2.0
werkzeug==3.0.1
tenacity==8.2.3
python-dotenv==1.0.0
